requirements.txt
Flask==1.1.2
future==0.18.2
lark-parser==0.9.0
torch==1.5.1
tqdm==4.48.0
